requirements_security.txt
# Add to requirements.txt
python-jose[cryptography]
passlib[bcrypt]
slowapi
